Checkout (temporary base) branches in detached state
When recreating a staging branch, the base branch needs to be checked
out. However, this fails when using multiple worktrees.

In these cases, check them out in detached state. This doesn't affect
the final staging branch in any way, since the tag is assigned later
and the branch is only used to prepare the worktree (and allow removal
of the existing ref)

Thanks to Etienne Laurin for the patch and initial test case

See MR !3